Move floor conversion functions to common_helpers.c
