+++ Controller/controller_helpers.c
@@ -12,6 +12,7 @@
 #include <arpa/inet.h>
 // my functions
 #include "controller_helpers.h"
+#include "../Car/car_helpers.h" // for the floor int char conversion functions
 
 int create_server()
 {
@@ -76,6 +77,4 @@ void handle_received_call_message(char *message)
   char source_floor[4];
   char destination_floor[4];
   sscanf(message, "%*s %3s %3s", source_floor, destination_floor);
-
-  printf("Call is from %s and is going to %s\n", source_floor, destination_floor);
 }
