+++ common_helpers.c
@@ -1,0 +1,23 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+int floor_char_to_int(char *floor)
+{
+  if (floor[0] == 'B')
+  {
+    floor[0] = '-';
+  }
+  return atoi(floor);
+}
+
+void floor_int_to_char(int floor, char *floorChar)
+{
+  if (floor < 0)
+  {
+    sprintf(floorChar, "B%d", abs(floor));
+  }
+  else
+  {
+    sprintf(floorChar, "%d", floor);
+  }
+}
