+++ common_helpers.h
@@ -1,0 +1,2 @@
+int floor_char_to_int(char *floor);
+void floor_int_to_char(int floor, char *floorChar);
